adds logic for overwriting previous data sources
if a company data source is to be created, we now check if one already exists for the company. if it does, a new company data source is created, and the original company data source date range is updated to expire at the same time as the new one's start.

this method allows us to keep historical snapshots of a company's data
sources.